requirements.txt
torch>=2.0.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=9.0.0
matplotlib>=3.5.0
openai>=1.0.0
python-dotenv>=0.19.0
pydantic>=1.10.0,<2.0.0
requests>=2.26.0
# meshy-api>=0.1.0
# sam2-hiera>=0.1.0
google-generativeai>=0.3.2
timm>=0.9.0
einops>=0.6.0
objaverse
transformers>=4.30.0
accelerate>=0.20.0
pybullet
pybullet_data
sentence-transformers
networkx>=3.0
matplotlib>=3.5.0